Adds inhumane schedule
