+++ pkg/controller/releasemanager/schedule/always.go
@@ -1,0 +1,7 @@
+package schedule
+
+import "time"
+
+func alwaysPermitsRelease(_ time.Time) bool {
+	return true
+}
